move propagation control tests
